+++ priv/repo/migrations/20190118235400_add_location_transactions_table.exs
@@ -5,8 +5,8 @@ defmodule BlockchainAPI.Repo.Migrations.AddLocationTransactionsTable do
     create table(:location_transactions) do
       add :owner, :string, null: false
       add :location, :string, null: false
-      add :nonce, :integer, null: false, default: 0
-      add :fee, :integer, null: false
+      add :nonce, :bigint, null: false, default: 0
+      add :fee, :bigint, null: false
 
       add :hash, references(:transactions, on_delete: :nothing, column: :hash, type: :string), null: false
       add :gateway, references(:gateway_transactions, on_delete: :nothing, column: :gateway, type: :string), null: false
